+++ dispatch/core/Exception/Exception.hh
@@ -39,6 +39,15 @@
 	std::this_thread::sleep_for(std::chrono::milliseconds(500));\
 	throw ex(ex_str.str(), __FILE__, __LINE__);\
 }
+
+#define CLAID_LOG_THROW_FATAL(ex, msg)\
+{\
+	std::ostringstream ex_str;\
+	ex_str << msg;\
+	Logger::logFatal("%s", ex_str.str().c_str());\
+	std::this_thread::sleep_for(std::chrono::milliseconds(500));\
+	throw ex(ex_str.str(), __FILE__, __LINE__);\
+}
 // std::this_thread::sleep_for(std::chrono::milliseconds(1000));
 
 namespace claid